requirements.txt
cma==3.3.0
gym==0.26.2
ipython==8.10.0
keras==2.11.0
matplotlib==3.7.0
numpy==1.24.2
tensorflow-macos==2.11.0
tqdm
